Use the size_t modifier to avoid an unnecessary cast
(cherry picked from commit 8f49a55c03ea8a2340f56124567bbe610906df9d)

+++ modules/gflags/gflags.c
@@ -157,7 +157,7 @@ static int fixup_gflags(void** param)
 
 	if ( myint >= 8*sizeof(*gflags) ) {
 		LM_ERR("flag <%d> out of "
-			"range [0..%lu]\n", myint, ((unsigned long)8*sizeof(*gflags))-1 );
+			"range [0..%zu]\n", myint, 8*sizeof(*gflags)-1);
 		return E_CFG;
 	}
 	/* convert from flag index to flag bitmap */
